% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gencline.R
\name{est_genocl}
\alias{est_genocl}
\title{Estimate genomic clines using a hierarchical model with unknown cline variances}
\usage{
est_genocl(
  Gx = NULL,
  G0 = NULL,
  G1 = NULL,
  p0 = NULL,
  p1 = NULL,
  H = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL,
  hier = TRUE,
  SDc = NULL,
  SDv = NULL,
  n_chains = 4
)
}
\arguments{
\item{Gx}{genetic data for putative hybrids, matrix for known genotypes, list of matrixes for genotype likelihoods, matrix of ancestry for the ancestry model}

\item{G0}{genetic data for parental reference set 0}

\item{G1}{genetic data for parental reference set 1}

\item{p0}{allele frequencies for parental reference set 0}

\item{p1}{allele frequencies for parental reference set 1}

\item{H}{hybrid indexes for the putative hybrids}

\item{model}{genotype, glik, or ancestry}

\item{ploidy}{diploid or mixed}

\item{pldat}{ploidy data for mixed ploidy}

\item{hier}{boolean, fit hierarchical model (TRUE) or assume cline SDs known (FALSE)}

\item{SDc}{known cline center SD on logit scale}

\item{SDv}{known cline gradient SD on log10 scale}

\item{n_chains}{number of HMC chains for posterior inference}
}
\value{
list of parameter estimates and full HMC results from stan, this includes cline parameters (center and gradient) and standard deviations describing variability in clines across loci (SDc and SDv)
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference of genomic clines from genetic data. This fits a hierarchical log-logistic model for genomic clines with two key parameters, cline center and cline gradient (i.e., slope, inversely proportional to cline width). The model also estimates the cline SDs across loci, SDc (variation in logit centers) and SDv (variation in log10 gradients).
}
