% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_geocline.R
\name{est_geocl}
\alias{est_geocl}
\title{Function to estimate geographic clines for a set of genetic loci or hybrid index}
\usage{
est_geocl(
  G = NULL,
  Geo = NULL,
  Ids = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL
)
}
\arguments{
\item{model}{genotype, glik, or ancestry}

\item{ploidy}{diploid or mixed}

\item{pldat}{ploidy data for mixed ploidy}

\item{Gx}{genetic data for putative hybrids}

\item{G0}{genetic data for parental reference set 0}

\item{G1}{genetic data for parental reference set 1}

\item{p0}{allele frequencies for parental reference set 0}

\item{p1}{allele frequencies for parental reference set 1}
}
\value{
list of parameter estimates and full HMC results from stan
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference of geographic clines from genetic data.
}
