% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_hi.R
\name{est_hi}
\alias{est_hi}
\title{Function to estimate hybrid index}
\usage{
est_hi(
  Gx = NULL,
  G0 = NULL,
  G1 = NULL,
  p0 = NULL,
  p1 = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL,
  n_chains = 4,
  n_iters = 2000,
  p_warmup = 0.5,
  n_thin = 1,
  n_cores = NULL
)
}
\arguments{
\item{Gx}{genetic data for putative hybrids in the form of a matrix for known genotypes (rows = individuals, columns = loci), a list of matrixes for genotype likelihoods (same dimensions but one matrix per genotype), or  matrix of ancestry for the ancestry model (rows = individuals, columns = loci).}

\item{G0}{genetic data for parental reference set 0 formatted as described for Gx.}

\item{G1}{genetic data for parental reference set 1 formatted as described for Gx.}

\item{p0}{vector of allele frequencies for parental reference set 0 (one entry per locus).}

\item{p1}{vector allele frequencies for parental reference set 1 (one entry per locus).}

\item{model}{for genetic data, either 'genotype' for known gentoypes, 'glik' for genotype likelihoods, or 'ancestry' for known ancestry.}

\item{ploidy}{species ploidy, either all 'diploid' or 'mixed' for diploid and haploid loci or individuals.}

\item{pldat}{matrix or list of matrixes of ploidy data for mixed ploidy (rows = individuals, columns = loci) indicating ploidy (2 = diploid, 1 = haploid).}

\item{n_chains}{number of HMC chains for posterior inference.}

\item{n_iters}{A positive integer specifying the number of iterations for each chain (including warmup), default is 2000.}

\item{p_warmup}{proportion (between 0 and 1) of n_iters to use as warmup (i.e., burnin), default is 0.5.}

\item{n_thin}{positive integer, save every n_thin HMC iterations for the posterior, default is 1.}

\item{n_cores}{number of cores to use for HMC, leave as NULL to automatically detect the number of cores present (no more than n_chains cores can be used even if available).}

\item{H}{vector of hybrid indexes for the putative hybrids (one entry per locus).}
}
\value{
A list of parameter estimates (hi = hybrid indexes) and full HMC results from stan. Parameter estimates are provided as a point estimate (median of the posterior) and 95\% equal-tail probability intervals (2.5th and 97.5th quantiles of the posterior distribution). These are provided as a vector or matrix depending on the dimensionality of the parameter. The full HMC output from rstan is provided as the final element in the list. This can be used for HMC diagnostics and to extract other model outputs not provided by default.
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference of hybrid indexes from genetic data.
}
\references{
Gompert Z, et al. 2024. Bayesian hybrid zone analyses with Hamiltonian Monte Carlo in R. Manuscript in preparation
}
\seealso{
'rstan::stan' for details on HMC with stan and the rstan HMC output object.
}
