% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_hi_vi.R
\name{est_hi_vi}
\alias{est_hi_vi}
\title{Function to estimate hybrid index using a variational algorithm for approximate posterior sampling}
\usage{
est_hi_vi(
  Gx = NULL,
  G0 = NULL,
  G1 = NULL,
  p0 = NULL,
  p1 = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL,
  method = "meanfield"
)
}
\arguments{
\item{Gx}{genetic data for putative hybrids in the form of a matrix for known genotypes (rows = individuals, columns = loci), a list of matrixes for genotype likelihoods (same dimensions but one matrix per genotype), or  matrix of ancestry for the ancestry model (rows = individuals, columns = loci).}

\item{G0}{genetic data for parental reference set 0 formatted as described for Gx.}

\item{G1}{genetic data for parental reference set 1 formatted as described for Gx.}

\item{p0}{vector of allele frequencies for parental reference set 0 (one entry per locus).}

\item{p1}{vector of allele frequencies for parental reference set 1 (one entry per locus).}

\item{model}{for genetic data, either 'genotype' for known genotypes, 'glik' for genotype likelihoods, or 'ancestry' for known ancestry.}

\item{ploidy}{species ploidy, either all 'diploid' or 'mixed' for diploid and haploid loci or individuals.}

\item{pldat}{matrix or list of matrixes of ploidy data for mixed ploidy (rows = individuals, columns = loci) indicating ploidy (2 = diploid, 1 = haploid).}

\item{method}{variational inference algorithm to use, either 'meanfield’ or 'fullrank’, default is meanfiled; fullrank deals better with covariances among parameters in the (approximate and transformed) posterior distribution}
}
\value{
A list of parameter estimates (hi = hybrid indexes). Parameter estimates are provided as a point estimate (median of the posterior), 90\% equal-tail probability intervals (5th and 95th quantiles of the posterior distribution), 95\% equal-tail probability intervals (2.5th and 97.5th quantiles of the posterior distribution). These are provided as a vector or matrix depending on the dimensionality of the parameter.
}
\description{
Uses Stan's variational algorithm for approximate posterior sampling for Bayesian inference of hybrid indexes from genetic data. This is an approximation to the full Bayesian analysis with HMC. This is an experimental feature and might not perform well in all cases, but will generally be much faster than HMC.
}
\details{
Hybrid indexes can be estimated from known genotypes (model = 'genotype'), genotype likelihoods (model = 'glik'), or known (estimated) ancestry (model = 'ancestry'). Genotypes should be encoded as 0 (homozygote), 1 (heterozygote) and 2 (alternative homozygote). No specific polarization (e.g., minor allele, reference allele, etc.) of 0 vs 2 is required. For haploid loci, use 0 and 1. Genotype likelihoods should be on their natural scale (not phred scaled) and the values for each locus for an individual should sum to 1. The data should be provided as a list of three matrixes, with the matrixes giving the likelihoods for genotypes 0, 1 and 2 respectively. Thus, each matrix will have one row per individual and one column per locus. For haploid loci with genotype likelihoods, you must use the 0 and 1 matrixes to store the likelihoods of the two possible states. For the ancestry model, hybrid indexes are inferred directly from known local (locus-specific) ancestry rather than from genotype data. Users are free to use whatever software they prefer for local ancestry inference (many exist). In this case, each entry in the individual (rows) by locus (columns) matrix should denote the number of gene copies inherited from parental population 1 (where pure parent 1 corresponds with a hybrid index of 1 and pure parent 0 corresponds with a hybrid index of 0). Haploids can be encoded using 0 and 1. For all models, missing data can be encoded by setting the ploidy for an individual/locus to 0 (this indicates no information, whereas genotype likelihoods encode uncertainty in genotypes) and the genotype to NA.

Hybrid genetic (or ancestry) data are always required. For genotype or genotype likelihood models, users must either provide pre-estimated parental allele frequencies or parent genetic (genotypes or genotype likelihoods) that can be used to infer allele frequencies. Parental data are not required for the ancestry model.

Ploidy data are only required for the mixed ploidy data. In this case, there should be one matrix for the hybrids or a list of matrixes for the hybrids (1st matrix) and each parent (2nd and 3rd matrixes, with parent 0 first). The latter is required for the genotype or genotype likelihood models if parental allele frequencies are not provided. The matrixes indicate whether each locus (column) for each individual (row) is diploid (2) or haploid (1) (use 0 for missing data).
}
\examples{
\dontrun{
## load the data set
data(genotypes)
## this includes three objects, GenHybrids, GenP0, and GenP1

## estimate parental allele frequencies, uses default HMC settings
p_out<-est_p(G0=GenP0,G1=GenP1,model="genotype",ploidy="diploid")

## estimate hybrid indexes, uses default HMC settings
## and uses point estimates (posterior medians) of allele frequencies
h_out<-est_hi_vi(Gx=GenHybrids,p0=p_out$p0[,1],p1=p_out$p1[,1],model="genotype",ploidy="diploid")
}
}
\references{
Gompert Z, DeRaad D, Buerkle CA. 2024. A next generation of hierarchical Bayesian analyses of hybrid zones enables model-based quantification of variation in introgression in R. Ecology and Evolution, 14:e70584.

Kucukelbir A, Tran D, Ranganath R, Gelman A, Blei DM. 2017. Automatic differentiation variational inference. Journal of Machine Learning Research, 18:1-45.
}
\seealso{
'rstan::vb' for details on variational inference with stan.
}
