% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_p.R
\name{est_p}
\alias{est_p}
\title{Function to estimate parental allele frequencies}
\usage{
est_p(
  G0 = NULL,
  G1 = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL
)
}
\arguments{
\item{G0}{genetic data for parental reference set 0}

\item{G1}{genetic data for parental reference set 1}

\item{model}{genotype, glik, or ancestry}

\item{ploidy}{diploid or mixed}

\item{pldat}{ploidy data for mixed ploidy}

\item{Gx}{genetic data for putative hybrids}
}
\value{
list of parameter estimates and full HMC results from stan
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference parental allele frequencies from genetic data.
}
