% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_Q.R
\name{est_Q}
\alias{est_Q}
\title{Function to estimate ancestry class proportions (Q)}
\usage{
est_Q(
  Gx = NULL,
  G0 = NULL,
  G1 = NULL,
  p0 = NULL,
  p1 = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL
)
}
\arguments{
\item{Gx}{genetic data for putative hybrids}

\item{G0}{genetic data for parental reference set 0}

\item{G1}{genetic data for parental reference set 1}

\item{p0}{allele frequencies for parental reference set 0}

\item{p1}{allele frequencies for parental reference set 1}

\item{model}{genotype, glik, or ancestry}

\item{ploidy}{diploid or mixed}

\item{pldat}{ploidy data for mixed ploidy}
}
\value{
list of parameter estimates and full HMC results from stan, this includes Q (ancestry class proportions) and hybrid index, which is derived from Q
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference of ancestry classes from genetic data. Ancestry classes denote the proportion of an individual's genome where both gene copies come from source 1 (Q11), both gene copies come from source 0 (Q00), or where one gene copy comes from source 1 and one from source 0.
}
