% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencline_plot.R
\name{gencline_plot}
\alias{gencline_plot}
\title{Plots genomic clines for a set of loci}
\usage{
gencline_plot(center = NULL, v = NULL, pdf = TRUE, outf = "tri_plot.pdf", ...)
}
\arguments{
\item{center}{vector of cline centers (from est_gencline.R)}

\item{v}{vector of cline gradients (from est_gencline.R)}

\item{pdf}{a logical specifying whether results should be output to a pdf file}

\item{outf}{a character string specifying the name of the output file if 'pdf=TRUE'}

\item{...}{additional arguments for plotting}
}
\value{
A plot is produced, but there is no return value
}
\description{
Plots a set of genomic clines.
}
