% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_plot.R
\name{pp_plot}
\alias{pp_plot}
\title{Plots bivariate posterior distribution of hierarchical cline parameters}
\usage{
pp_plot(
  objs = NULL,
  param1 = "muc",
  param2 = "sdc",
  probs = c(0.5, 0.75, 0.95),
  colors = "black",
  addPoints = TRUE,
  palpha = 0.3,
  pdf = TRUE,
  outf = "pp_plot.pdf",
  ...
)
}
\arguments{
\item{objs}{the output from est_gencline, or a list of such objects}

\item{param1}{a string specifying the genomic cline parameter (muc, muv, sc or sv) for the x-axis on the plot.}

\item{param2}{a string specifying the genomic cline parameter (muc, muv, sc or sv) for the y-axis on the plot.}

\item{probs}{vector specifying which confidence ellipses to plot (e.g., 0.95 for an ellipse that encloses 95\% of the posterior).}

\item{colors}{single color or vector of colors for plotting, if more than one color is supplied different colors will be used for different posteriors (in the order specified by objs).}

\item{addPoints}{Boolean indicating whether points denoting samples from the posterior should be added to the plot (default = TRUE).}

\item{palpha}{transparency (alpha) value for the color of points, must be between 0 and 1, smaller values make points more transparent.}

\item{pdf}{a logical specifying whether results should be output to a pdf file; if false the plot is sent to the default graphics device.}

\item{outf}{a character string specifying the name of the output file if 'pdf=TRUE' \link{default = tri_plot.pdf}.}

\item{...}{additional arguments for plotting, see options in par and plot.}
}
\value{
A plot is produced, but there is no return value.
}
\description{
Creates bivariate plots of combinations of cline standard deviations and means for one or multiple sets of loci.
}
\details{
This function produces a plot illustrating the bivariate posterior probability distribution for the cline mean or standard deviation parameters for one or more sets of genetic loci. Bivariate plots can be produced for any combination of higher-level cline parameters, that is the mean center (muc), mean slope (muv), standard deviation of centers (sc) or standard deviation of slopes (sv) (these are the names of the parameters in the HMC objects and should be used for param1 and param2). Plots include confidence ellipses cpaturing user specified proportions of the posterior (set with probs) and optionally points representing the samples from the posterior. Ellipses are produced by approximating the bivariate posterior with a bivariate normal distribution (these are meant to serve an illustrative purpose only). Posteriors from multiple sets of loci (multiple outputs from the est_gencline function) can be placed on the same plot. To plot a single posterior, pass the entire object (a list) produced by est_gencline to the function for the objs argument. To plot multiple, pass a list of these objects (a list of lists) to the objs function.

Users can specify colors for plotting posteriors. When more than one color is provided, different colors will be used for each object (set of loci). The transparency of the points (if included) can be adjusted with the palpha parameter. Transperancy of the ellipses is set automatically by setting the alpha (transparency) value for each ellipses to 1-probs.
}
