% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum2zero.R
\name{sum2zero}
\alias{sum2zero}
\title{Impose hard sum-to-zero constraints on cline estimates}
\usage{
sum2zero(center = NULL, v = NULL, hmc = NULL, transform = TRUE, ci = 0.95)
}
\arguments{
\item{center}{vector or matrix of cline centers (from est_gencline). If a vector, there should be one element per locus; if a matrix, there should be one row per loucs and one column each for the point estimate and lower and upper bounds of the credible interval.}

\item{v}{vector or matrix of cline gradients (from est_gencline). If a vector, there should be one element per locus; if a matrix, there should be one row per locus and one column each for the point estimate and lower and upper bounds of the credible interval.}

\item{hmc}{HMC object from est_gencline.}

\item{transform}{Boolean variable indicating whether to apply the constraint on the natural scale of v and center (FALSE) or on the transformed scale of log(v) and logit(center). We recommend applying these transformations.}

\item{ci}{size of the credible interval, specifically, the equal-tail probability interval, to generate from the HMC object (if supplied) \link{default = 0.95}.}
}
\value{
A list with two vectors (if only a parameter estimate vector was provided) or a matrixes (if a matrix of the full HCM object was given) with the re-calculated, constrained parameter estimates. If the HMC object was given, a point estimate and the bounds of hte specified credible intervals are given.
}
\description{
Re-calculates cline parameter estimates to ensure that the average cline parameter corresponds with expectations for genome-average introgression.
}
\details{
Genomic clines model locus-specifc patterns of introgression relative to genome-average introgression or ancestry. Thus, if the same loci are used to estimate hybrid index and genomic clines (or if the former are a random sample of the latter), we would expect the average deviations from genome-average admixture to be 0 across loci (i.e., the deviations should cancel out). This is suggested by prior structure of the hierarchical Bayesian model, where the mean for log10(v) and logit(center) are set to 0. This is a soft sum-to-zero constraint (the prior pulls the sum towards zero, but a sum-to-zero constraint is not enforced). This function instead enforced a hard sum-to-zero constraint. This is done either by working with the full HMC output or parameter estimates (just point estimates or point estimates and credible intervals). In the former case, the mean cline parameters (on the log10 or logit scale, as appropriate) at each HMC iteration are forced to be 0 by subtracting off the mean. Point estimates and credible intervals are then re-calculated. In the case of parameter estimates, the mean point estimate (on the log10 or logit scale) is subtracted from the point estimate and credible intervals. Using the full HMC object is generally preferable.

As an alternative, the constraint can be applied on the natural scale rather than the log or logit scale, that is, the mean cline center can be constrained to 0.5 and the mean gradient (v) to 1. For center, the difference will often be trivial. For gradient the difference could be greater, and my current suggestion is to use the log10 scale as v is a ratio.
}
\references{
Gompert Z, DeRaad D, Buerkle CA. A next generation of hierarchical Bayesian analyses of hybrid zones enables model-based quantification of variation in introgression in R. bioRxiv 2024.03.29.587395.
}
\seealso{
'rstan::stan' for details on HMC with stan and the rstan HMC output object.
}
