% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgchm-package.R
\docType{package}
\name{bgchm-package}
\alias{bgchm-package}
\alias{bgchm}
\title{The 'bgchm' package.}
\description{
This is an R package for Bayesian analyses of population genomic data from hybrid zones, including Bayesian genomic cline analysis, estimation of hybrid indexes and ancestry class proportions, some geographic cline analyses, and accessory plotting functions. This package using Hamiltonian Monte Carlo (HMC) for sampling posterior distributions, with HMC sampling implemented via Stan.
}
\references{
Gompert Z, et al. 2024. Bayesian analyses of hybrid zones in R with Hamiltonian Monte Carlo. Manuscript in preparation.
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.7. https://mc-stan.org
}
