% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgchm-package.R
\docType{package}
\name{bgchm-package}
\alias{bgchm-package}
\alias{bgchm}
\title{The 'bgchm' package.}
\description{
This is an R package for Bayesian analyses of population genomic data from hybrid zones, including Bayesian genomic cline analysis, estimation of hybrid indexes and ancestry class proportions, some geographic cline analyses, and accessory plotting functions. This package using Hamiltonian Monte Carlo (HMC) for sampling posterior distributions, with HMC sampling implemented via Stan.
}
\references{
Gompert Z, DeRaad D, Buerkle CA. 2024. A next generation of hierarchical Bayesian analyses of hybrid zones enables model-based quantification of variation in introgression in R. Ecology and Evolution, 14:e70584.
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.7. https://mc-stan.org
}
\author{
\strong{Maintainer}: Zachariah Gompert \email{zach.gompert@usu.edu} (\href{https://orcid.org/0000-0003-2248-2488}{ORCID})

}
