% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pp.R
\name{calc_pp}
\alias{calc_pp}
\title{Function to calcualte posterior probabilities of extreme cline parameter values}
\usage{
calc_pp(hmc = NULL, param = NULL, greater = TRUE, refval = NULL)
}
\arguments{
\item{hmc}{list object produced by est_genocl or est_geocl.}

\item{param}{character or string specifying the cline parameter to summarize.}

\item{greater}{Boolean indicating whether to compute the posterior probability the value is greater than (TRUE) or less than (FALSE) the null or reference value.}

\item{refval}{null or reference value for posterior probability calculation.}
}
\value{
A vector of posterior probabilities (of length > 1 for vector parameters, such as cline slopes with multiple loci).
}
\description{
Computes the posterior probability that cline parameter values exceed a user defined null or reference value.
}
\details{
This function computes the posterior probability the a parameter value exceeds (or is less than) a null or reference value based on samples from the posterior distribution. This works with geographic or genomic cline output and for any model parameter in the HMC output based on the name of the parameter in the HMC object (i.e., Stan model). For genomic clines, this includes cline centers (center), slopes (v), cline standard deviations (sc and sv) and cline means (muc and muv) (not all parameters are components of all models). For geographic clines, this includes  cline centers (cent), slopes (slope), cline widths (w), the mean cline width (mu), and the standard deviation in cline slopes (sigma). In both cases, the name of the parameter in the HMC object is given in parantheses and should be provided as a character or string to the param argument. Different reference or null values are relevant for different questions or parameters. For example, the expected average genomic cline slope, v, is 1, and thus one could identify loci with steeper (or shallower) genomic clines than the genome-average by computing the posterior probability that the slope, v, exceeds (or is less than) 1.
}
\references{
Gompert Z, et al. 2024. Bayesian hybrid zone analyses with Hamiltonian Monte Carlo in R. Manuscript in preparation
}
\seealso{
'rstan::stan' for details on HMC with stan and the rstan HMC output object.
}
