% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gencline_vi.R
\name{est_genocl_vi}
\alias{est_genocl_vi}
\title{Estimate genomic clines using Stan's variational algorithm for approximate posterior sampling}
\usage{
est_genocl_vi(
  Gx = NULL,
  G0 = NULL,
  G1 = NULL,
  p0 = NULL,
  p1 = NULL,
  H = NULL,
  SDc = 100,
  SDv = 100
)
}
\arguments{
\item{Gx}{genetic data for putative hybrids in the form of a matrix for known genotypes (rows = individuals, columns = loci), a list of matrixes for genotype likelihoods (same dimensions but one matrix per genotype), or  matrix of ancestry for the ancestry model (rows = individuals, columns = loci).}

\item{G0}{genetic data for parental reference set 0 formatted as described for Gx.}

\item{G1}{genetic data for parental reference set 1 formatted as described for Gx.}

\item{p0}{vector of allele frequencies for parental reference set 0 (one entry per locus).}

\item{p1}{vector allele frequencies for parental reference set 1 (one entry per locus).}

\item{H}{vector of hybrid indexes for the putative hybrids (one entry per individual).}

\item{SDc}{known cline center SD on logit scale.}

\item{SDv}{known cline gradient SD on log10 scale.}
}
\value{
A list of parameter estimates, specifically the cline parameters (center and gradient). Parameter estimates are provided as a point estimate (median of the posterior) and 90\% equal-tail probability intervals (5th and 95th quantiles of the posterior distribution). These are provided as a vector or matrix depending on the dimensionality of the parameter.
}
\description{
Estimates genomic clines from genetic data using Stan's variational algorithm for approximate posterior sampling. This is currently an experimental feature that should be used with caution. At present, this function only works with diploids and known genotypes and standard deviation parameters (e.g., pre-estimated from the hierarchical model with HMC).
}
\details{
At present, this function only fits clines based on known genotypes and diploid loci (I will introduce additional options after further testing). Genotypes should be encoded as 0 (homozygote), 1 (heterozygote) and 2 (alternative homozygote). No specific polarization (e.g., minor allele, reference allele, etc.) of 0 vs 2 is required. Hybrid indexes must be provided. Genetic  data for putative hybrids are also always required. Users must either provide pre-estimated parental allele frequencies or parent genotypes that can be used to infer allele frequencies with est_p. Cline SD parameters must be proived. This can be SD parameters estimated from an initital hierarchical model with HMC or large (weakly informative) values if one does not want to approximate a hierarchical model.
}
\references{
Gompert Z, DeRaad D, Buerkle CA. A next generation of hierarchical Bayesian analyses of hybrid zones enables model-based quantification of variation in introgression in R. bioRxiv 2024.03.29.587395.
}
\seealso{
'rstan::vb' for details on variational inference with stan.
}
