% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_geocline.R
\name{est_geocl}
\alias{est_geocl}
\title{Function to estimate geographic clines for a set of genetic loci or hybrid index}
\usage{
est_geocl(
  G = NULL,
  P = NULL,
  Geo = NULL,
  Ids = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL,
  prec = 0.001
)
}
\arguments{
\item{G}{genetic data}

\item{P}{allele frequency data}

\item{Geo}{geographic coordinates for the sampled populations, these will be centered automatically}

\item{Ids}{indexes designating which population each individual belongs to, indexes run from 1 to the number of populations}

\item{model}{genotype, glik, or ancestry}

\item{ploidy}{diploid or mixed}

\item{pldat}{ploidy data for mixed ploidy}

\item{prec}{precision for known allele frequencies, set to approximately 1/2N}
}
\value{
list of parameter estimates and full HMC results from stan
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference of geographic clines from genetic data.
}
