% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_hi.R
\name{est_hi}
\alias{est_hi}
\title{Function to estimate hybrid index}
\usage{
est_hi(
  Gx = NULL,
  G0 = NULL,
  G1 = NULL,
  p0 = NULL,
  p1 = NULL,
  model = "genotype",
  ploidy = "diploid",
  pldat = NULL
)
}
\arguments{
\item{Gx}{genetic data for putative hybrids}

\item{G0}{genetic data for parental reference set 0}

\item{G1}{genetic data for parental reference set 1}

\item{p0}{allele frequencies for parental reference set 0}

\item{p1}{allele frequencies for parental reference set 1}

\item{model}{genotype, glik, or ancestry}

\item{ploidy}{diploid or mixed}

\item{pldat}{ploidy data for mixed ploidy}
}
\value{
list of parameter estimates and full HMC results from stan
}
\description{
Uses Hamiltonian Monte Carlo (HMC) for Bayesian inference of hybrid index from genetic data.
}
