% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_plot.R
\name{tri_plot}
\alias{tri_plot}
\title{Plots interpopulation ancestry (Q10) as a function of hybrid index}
\usage{
tri_plot(hi = NULL, Q10 = NULL, pdf = TRUE, outf = "tri_plot.pdf", ...)
}
\arguments{
\item{hi}{a vector of hybrid index estimates (from est_h or est_Q)}

\item{Q10}{a vector of interpopulation ancestry estimates (from est_Q)}

\item{pdf}{a logical specifying whether results should be output to a pdf file}

\item{outf}{a character string specifying the name of the output file if 'pdf=TRUE'}

\item{...}{additional arguments for plotting}
}
\value{
A plot is produced, but there is no return value
}
\description{
Creates a triangle plot of hybrid index versus interpopulation ancestry.
}
